+++ packages/account-api-sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sophon-labs/account-api-sdk
 
+## 2.1.5
+
+### Patch Changes
+
+- Changing initialization order on rn
+- Updated dependencies
+  - @sophon-labs/account-core@2.1.5
+
+## 2.1.4
+
+### Patch Changes
+
+- Cleanup on authentication on fresh install
+- Updated dependencies
+  - @sophon-labs/account-core@2.1.4
+
+## 2.1.3
+
+### Patch Changes
+
+- Removing disconnect from react native and adding is initialized
+- Updated dependencies
+  - @sophon-labs/account-core@2.1.3
+
 ## 2.1.2
 
 ### Patch Changes